% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DaO.R
\name{DaO_calloption}
\alias{DaO_calloption}
\title{DaO-calloption: Down and Out Call Option}
\usage{
DaO_calloption(t, mu, sigma, K, B, S0 = 1)
}
\arguments{
\item{t}{time to maturity in day (integer)}

\item{mu}{drift (float)}

\item{sigma}{volatility (float)}

\item{K}{strike price (float)}

\item{B}{barrier price}

\item{S0}{value of stock at the end of the zero day (float)}
}
\value{
v: float: value of option at maturity
}
\description{
Compute the value of the option at maturity
}
\examples{
res <- DaO_calloption(100, 0.05, 0.005, 2, 0.2)
}
